ReviveXR: work around bad gamma handling in WMR
WMR always sends linear colors to displays that should have gamma
correction applied, and in fact will decode sRGB colors into
linear space before sending them to a display that demands a
gamma correction. This is incorrect, so to work around this we
make the DX render target sRGB but tell WMR it's linear.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>