Update multiplayer fix (#1573)
* ReviveOverlay: Fix heartbeats not properly enabling multiplayer

The Oculus API is now signed into on app launch with the titles appID, rather than on dashboard launch with the Oculus Platform appID

* Appveyor: Update OpenSSL DLL's to v1.1.1

OpenSSL renamed their DLL's in a newer release. Copying both sets of DLL's just to be safe, as this can prevent the dashboard from properly obtaining game titles for manifest generation, as well as sending heartbeats.